SingleAssignmentGrid component add, Assignment.jsx update to show SingleAssignmentGrid
